fix: fixed code scanning alerts (#49)
Fix code scanning alert - Comparison of narrow type with wide type in loop condition #42

+++ src/main/java/org/outerj/daisy/diff/html/HTMLDiffer.java
@@ -170,8 +170,8 @@ public class HTMLDiffer implements Differ {
 
             while (i + 1 < differences.length
                     && differences[i + 1].kind() == kind
-                    && score(leftLength, differences[i + 1].leftLength(),
-                    rightLength, differences[i + 1].rightLength()) > (differences[i + 1]
+                    && ((int) score(leftLength, differences[i + 1].leftLength(),
+                    rightLength, differences[i + 1].rightLength())) > (differences[i + 1]
                     .leftStart() - leftEnd)) {
                 leftEnd = differences[i + 1].leftEnd();
                 rightEnd = differences[i + 1].rightEnd();
